LayoutTests/imported/w3c/web-platform-tests/webcodecs/temporal-svc-encoding.https.any.worker_vp8-expected.txt


Harness Error (FAIL), message = Error in remote https://localhost:9443/resources/testharness.js: Error: assert_own_property: expected property "svc" missing

FAIL SVC L1T2 assert_equals: expected 12 but got 0
FAIL SVC L1T3 assert_equals: expected 6 but got 0
